+++ src/topBar/TopBar.unit.tsx
@@ -5,37 +5,19 @@ import { Button } from '../button'
 import { TopBar } from './TopBar'
 
 describe('TopBar', () => {
-  it('should not have any modifier classes', () => {
-    const topBar = shallow(<TopBar />)
-    expect(topBar.hasClass('kirk-topBar--fixed')).toBe(false)
-    expect(topBar.hasClass('kirk-topBar--overlayed')).toBe(false)
-    expect(topBar.hasClass('kirk-topBar--bgTransparent')).toBe(false)
-    expect(topBar.hasClass('kirk-topBar--bgShadedTransparent')).toBe(false)
+  it('should not have the scrolled class if hasScrolled is false or null', () => {
+    const topBar = shallow(<TopBar zIndex={50} />)
+    expect(topBar.hasClass('kirk-topBar--scrolled')).toBe(false)
   })
 
-  it('should have the custom class on inner wrapper', () => {
-    const topBar = shallow(<TopBar innerWrapperClassName="test" />)
-    expect(topBar.find('.test').exists()).toBe(true)
-  })
-
-  it('should have the fixed modifier class', () => {
-    const topBar = shallow(<TopBar fixed />)
-    expect(topBar.hasClass('kirk-topBar--fixed')).toBe(true)
-  })
-
-  it('should have the bgTransparent modifier class', () => {
-    const topBar = shallow(<TopBar bgTransparent />)
-    expect(topBar.hasClass('kirk-topBar--bgTransparent')).toBe(true)
-  })
-
-  it('should have the bgShadedTransparent modifier class', () => {
-    const topBar = shallow(<TopBar bgShadedTransparent />)
-    expect(topBar.hasClass('kirk-topBar--bgShadedTransparent')).toBe(true)
+  it('should not have the scrolled class if hasScrolled is true', () => {
+    const topBar = shallow(<TopBar zIndex={50} hasScrolled />)
+    expect(topBar.hasClass('kirk-topBar--scrolled')).toBe(true)
   })
 
   it('should have a clickable button ', () => {
     const onClick = jest.fn()
-    const topBar = mount(<TopBar leftItem={<Button onClick={onClick} />} />)
+    const topBar = mount(<TopBar zIndex={50} leftItem={<Button onClick={onClick} />} />)
     expect(topBar.find('button')).toHaveLength(1)
     const button = topBar.find('button')
     button.simulate('click')
@@ -43,19 +25,19 @@ describe('TopBar', () => {
   })
 
   it('should have a left area with a button', () => {
-    const topBar = mount(<TopBar leftItem={<Button />} />)
+    const topBar = mount(<TopBar zIndex={50} leftItem={<Button />} />)
     const leftArea = topBar.find('.kirk-topBar-left')
     expect(leftArea.contains(<Button />)).toBe(true)
   })
 
   it('should have a right area with a span', () => {
-    const topBar = mount(<TopBar rightItem={<span>Test</span>} />)
+    const topBar = mount(<TopBar zIndex={50} rightItem={<span>Test</span>} />)
     const rightArea = topBar.find('.kirk-topBar-right')
     expect(rightArea.contains(<span>Test</span>)).toBe(true)
   })
 
   it('should have a center area', () => {
-    const topBar = mount(<TopBar centerItem={<span>Test</span>} />)
+    const topBar = mount(<TopBar zIndex={50} centerItem={<span>Test</span>} />)
     expect(topBar.find('.kirk-topBar-center')).toHaveLength(1)
   })
 })